refactor: move playlist metadata orchestration to service owner
SeriesPlaylist previously imported three metadata services and owned their
asynchronous loading policy. SeriesPlaylistMetadataLoader now owns four static
operations and six private helpers. Exactly nine production calls migrate across
playlist content, player and launcher; existing continuation statements remain.

The model retains its final private per-playlist map. Typed cachedMovieImdbId
has no shared-ID fallback or bounds check. Synchronous recordMovieMetadataSuccess
writes the index cache BEFORE shared imdbId ??=. Preserve null-default parameter
versus existing field, valid supplied tt assignment before first await, mapping
precedence, cached-read-before-bounds, first successful completion, empty shared
retention and independent overlapping lookups. No dedup/new lifecycle/bootstrap.

Moved454 original lines; explicit receiver/static qualifiers and two cache API
operations replace private access. Existing await/catch/mutation order remains;
wrapper still awaits index0. Complete body/caller inverse ledger retained locally.
The exact35 test invocation substitutions plusone import reverse to original19
SHA0af94c25ab781611f21ce12bfec4f66f5a4ad8445d052a7c5793cb7ea4b7f422.
Before-move origin19 is separately frozen f19993226f402e0815c0ce8632e48f12b37d4b6d,
with e179/a3ceec origin history and initial7/1 then8/13 greens preserved.

First product scoped pass STOPPED: parent129/head129 diagnostics, +1/-1 because
unchanged emptycatch moved to a new path. Preserve firstRED rawSHA
580854dc895b3d54a6d988040206ed20994beb743916c1d08d353c1ced42b525.
Only approved explanatory comment added inside catch; following returnnull and
catchscope unchanged. No suppression/baseline waiver. Comment deltaSHA
56b9a486614df625020d7b7bb31536de22710aed03537dd248a65a5af6ec7492.

Pinned Flutter3.44.8, external180s bounds, independently accepted evidence:
- Reused parent scoped evidence; corrected129->128, +0/-1, inheritedexit2 both,
  zero errors and owner/test diagnostics.
- ONE focused35: adaptedmetadata19 +series_browser_guide6 +movie_resume_source10,
  flutter test --no-pub --reporter json;35PASS/0error,exit0,JSONtrue5048ms.
  RawSHA0d1a7ed695b41092f3859ce86deae6280f68acabfda0dd6b7c0124ce92a4565e.
- Actual py -3 tool/analyze_baseline.py once:430current/449baseline,0new,exit0.
- Same actual tool/check_layering.dart parent/head JSON and ci_layering_delta.py:
  56->53,+0/-3; Counter proves exactthree model-service import identities removed.
Evidence .dart_tool/series-playlist-metadata-origin/choice-b-validation/ includes
terminal-corrected-proof.json, raw logs and source hashes. A post-helper display
UTF8 decode failure was read-only decoded as CP1252; analyzer never rerun.

Exactly five production paths plus adaptedtest; production Git+491/-471 net+20 after
comment correction; test+36/-35 net+1. No CODEMAP, deps or baseline edits.
No native/device validation. Guide/resume suites do not execute allnine changed
caller continuations; their preservation is exactsourcehunk evidence only.
Four removed model methods are an explicit Dart source API migration, including
lost instance override dispatch; two new cache methods add public capabilities.
No current repository subclasses found; external API/subclass compatibility is
not promised. Threshold/ties/provider permutations/retry/expiry and complete
serialization remain finite gaps. Foundation model dependency remains; this is
three actual forbidden edges removed, not strict-Q1 closure.
