+++ lib/screens/playlist_content_view_screen.dart
@@ -1,3 +1,4 @@
+import '../services/series_playlist_metadata_loader.dart';
 import 'package:debrify/services/storage/playback_progress_store.dart';
 import 'package:debrify/services/storage/provider_credential_prefs.dart';
 import 'dart:async';
@@ -2219,8 +2220,7 @@ class _PlaylistContentViewScreenState extends State<PlaylistContentViewScreen> {
 
       // Fetch TVMaze metadata asynchronously
       if (_seriesPlaylist!.isSeries) {
-        _seriesPlaylist!
-            .fetchEpisodeInfo(
+        SeriesPlaylistMetadataLoader.fetchEpisodeInfo(_seriesPlaylist!,
               playlistItem: widget.playlistItem,
               imdbId: widget.playlistItem?['imdbId'] as String?,
             )
@@ -2316,8 +2316,7 @@ class _PlaylistContentViewScreenState extends State<PlaylistContentViewScreen> {
           _isLoadingSeriesMetadata = true;
         });
 
-        _seriesPlaylist!
-            .fetchEpisodeInfo(
+        SeriesPlaylistMetadataLoader.fetchEpisodeInfo(_seriesPlaylist!,
               playlistItem: widget.playlistItem,
               imdbId: widget.playlistItem?['imdbId'] as String?,
             )
