+++ lib/services/video_player_launcher.dart
@@ -1,3 +1,4 @@
+import 'series_playlist_metadata_loader.dart';
 import 'package:debrify/services/storage/quick_play_policy_prefs.dart';
 import 'package:debrify/services/storage/iptv_prefs.dart';
 import 'package:debrify/services/storage/playback_progress_store.dart';
@@ -2391,7 +2392,7 @@ class VideoPlayerLauncher {
           // Fetch TVMaze metadata so items arrive pre-populated with artwork/descriptions
           if (isSeries && seriesPlaylist != null) {
             try {
-              await seriesPlaylist.fetchEpisodeInfo(
+              await SeriesPlaylistMetadataLoader.fetchEpisodeInfo(seriesPlaylist,
                 imdbId: effectiveContentImdbId(),
               );
               debugPrint(
@@ -3472,7 +3473,7 @@ class VideoPlayerLauncher {
           debugPrint(
             'MovieAsync: Not a series, attempting movie metadata fetch',
           );
-          await seriesPlaylist.fetchMovieMetadata();
+          await SeriesPlaylistMetadataLoader.fetchMovieMetadata(seriesPlaylist);
 
           final discoveredImdbId = seriesPlaylist.imdbId;
           if (discoveredImdbId != null) {
@@ -3512,7 +3513,7 @@ class VideoPlayerLauncher {
         debugPrint(
           'TVMazeAsync: Calling fetchEpisodeInfo() with imdbId=$contentImdbId',
         );
-        await seriesPlaylist.fetchEpisodeInfo(imdbId: contentImdbId);
+        await SeriesPlaylistMetadataLoader.fetchEpisodeInfo(seriesPlaylist, imdbId: contentImdbId);
         debugPrint('TVMazeAsync: fetchEpisodeInfo() completed');
 
         // Save discovered IMDB ID back to playlist item for future direct plays
